Refactor: Fix issue with TwoPane mode when recipes are added and removed from favorites list.
The issue was with Ingredient and Instruction fragments onLoadFinished
getting triggered upon the DB record update was causing undesired
behavior. To simplify things, it is better to query all the required
data at beginning and then pass it along as a bundle to various
fragments or activities.

There also has been few work around to reset selected recipe item and
to adapt the layout in BakeliciousActivity in TwoPane mode.